refactor: don't re-define transitionStyles at element level
